Mini Project 1 Extension
Hello! I'm doubting that this is the correct way to post assignments, but hopefully this can be accessed! Definitely some issues with this code, but I tried a few things and got thinking a bit more in depth about this. Thank you for bearing with me--I look forward to your feedback!
-Caroline