Add files via upload
Please note that there's a small discrepancy between the format of the actual columns in the command line versus how this renders as a .csv file. I wasn't able to resolve this particular issue and have added more details in the readme file associated with this project. I look forward to your feedback and getting more clarification on this! 
-Caroline